Add Silk and wrap MassItemStorage in an Inventory implementation
